+++ cloudmanager/resource_netapp_cloudmanager_cifs_server.go
@@ -3,6 +3,7 @@ package cloudmanager
 import (
 	"fmt"
 	"log"
+	"strings"
 
 	"github.com/hashicorp/terraform/helper/schema"
 )
@@ -20,28 +21,28 @@ func resourceCVOCIFS() *schema.Resource {
 		Schema: map[string]*schema.Schema{
 			"domain": {
 				Type:     schema.TypeString,
-				Optional: true,
+				Required: true,
 				ForceNew: true,
 			},
 			"username": {
 				Type:     schema.TypeString,
-				Optional: true,
+				Required: true,
 				ForceNew: true,
 			},
 			"password": {
 				Type:      schema.TypeString,
-				Optional:  true,
+				Required:  true,
 				Sensitive: true,
 				ForceNew:  true,
 			},
 			"dns_domain": {
 				Type:     schema.TypeString,
-				Optional: true,
+				Required: true,
 				ForceNew: true,
 			},
 			"ip_addresses": {
-				Type:     schema.TypeSet,
-				Optional: true,
+				Type:     schema.TypeList,
+				Required: true,
 				ForceNew: true,
 				Elem: &schema.Schema{
 					Type: schema.TypeString,
@@ -49,12 +50,12 @@ func resourceCVOCIFS() *schema.Resource {
 			},
 			"netbios": {
 				Type:     schema.TypeString,
-				Optional: true,
+				Required: true,
 				ForceNew: true,
 			},
 			"organizational_unit": {
 				Type:     schema.TypeString,
-				Optional: true,
+				Required: true,
 				ForceNew: true,
 			},
 			"svm_name": {
@@ -81,17 +82,39 @@ func resourceCVOCIFS() *schema.Resource {
 				Type:     schema.TypeString,
 				Optional: true,
 				ForceNew: true,
+				ValidateFunc: func(val interface{}, key string) (warns []string, errs []error) {
+					v := val.(string)
+					if strings.TrimSpace(v) != "" {
+						errs = append(errs, fmt.Errorf("using workgroup configuration is deprecated. Create with AD instead"))
+					}
+					return
+				},
 			},
 			"workgroup_name": {
 				Type:     schema.TypeString,
 				Optional: true,
 				ForceNew: true,
+				ValidateFunc: func(val interface{}, key string) (warns []string, errs []error) {
+					v := val.(string)
+					if strings.TrimSpace(v) != "" {
+						errs = append(errs, fmt.Errorf("using workgroup configuration is deprecated. Create with AD instead"))
+					}
+					return
+				},
 			},
 			"is_workgroup": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				ForceNew: true,
-				Default:  false,
+				Type:       schema.TypeBool,
+				Optional:   true,
+				ForceNew:   true,
+				Default:    false,
+				Deprecated: "use AD instead",
+				ValidateFunc: func(val interface{}, key string) (warns []string, errs []error) {
+					v := val.(bool)
+					if v {
+						errs = append(errs, fmt.Errorf("using workgroup configuration is deprecated. Create with AD instead"))
+					}
+					return
+				},
 			},
 		},
 	}
@@ -102,56 +125,38 @@ func resourceCVOCIFSCreate(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*Client)
 	clientID := d.Get("client_id").(string)
 	cifs := cifsRequest{}
-	cifs.isWorkgroup = d.Get("is_workgroup").(bool)
-	if cifs.isWorkgroup {
-		if v, ok := d.GetOk("server_name"); ok {
-			cifs.ServerName = v.(string)
-		}
-		if v, ok := d.GetOk("workgroup_name"); ok {
-			cifs.WorkgroupName = v.(string)
-		}
-	} else {
-		if v, ok := d.GetOk("domain"); ok {
-			cifs.Domain = v.(string)
-		}
-		if v, ok := d.GetOk("username"); ok {
-			cifs.Username = v.(string)
-		}
-		if v, ok := d.GetOk("password"); ok {
-			cifs.Password = v.(string)
-		}
-		if v, ok := d.GetOk("dns_domain"); ok {
-			cifs.DNSDomain = v.(string)
-		}
-		if v, ok := d.GetOk("ip_addresses"); ok {
-			ips := make([]string, 0, v.(*schema.Set).Len())
-			for _, x := range v.(*schema.Set).List() {
-				ips = append(ips, x.(string))
-			}
-			cifs.IPAddresses = ips
-		}
-		if v, ok := d.GetOk("netbios"); ok {
-			cifs.NetBIOS = v.(string)
-		}
-		if v, ok := d.GetOk("organizational_unit"); ok {
-			cifs.OrganizationalUnit = v.(string)
-		}
+
+	cifs.Domain = d.Get("domain").(string)
+	cifs.Username = d.Get("username").(string)
+	cifs.Password = d.Get("password").(string)
+	cifs.DNSDomain = d.Get("dns_domain").(string)
+	IPAddresses := d.Get("ip_addresses")
+
+	for _, IPAddress := range IPAddresses.([]interface{}) {
+		cifs.IPAddresses = append(cifs.IPAddresses, IPAddress.(string))
 	}
+
+	cifs.NetBIOS = d.Get("netbios").(string)
+	cifs.OrganizationalUnit = d.Get("organizational_unit").(string)
 
 	workingEnvDetail, err := client.getWorkingEnvironmentDetail(d, clientID)
 	if err != nil {
 		return err
 	}
 	cifs.WorkingEnvironmentID = workingEnvDetail.PublicID
-	cifs.WorkingEnvironmentType = workingEnvDetail.WorkingEnvironmentType
-	cifs.SvmName = workingEnvDetail.SvmName
+
+	if v, ok := d.GetOk("svm_name"); ok {
+		cifs.SvmName = v.(string)
+	} else {
+		cifs.SvmName = workingEnvDetail.SvmName
+	}
 
 	err = client.createCIFS(cifs, clientID)
 	if err != nil {
 		log.Print("Error creating cifs")
 		return err
 	}
-	d.SetId(cifs.WorkingEnvironmentID)
+	d.SetId(cifs.SvmName)
 	return resourceCVOCIFSRead(d, meta)
 }
 
@@ -161,62 +166,58 @@ func resourceCVOCIFSRead(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*Client)
 	clientID := d.Get("client_id").(string)
 	cifs := cifsRequest{}
-	cifs.isWorkgroup = d.Get("is_workgroup").(bool)
-	// TO DO: get cifs-workgroup api missing.
-	if cifs.isWorkgroup {
-		return nil
+
+	if v, ok := d.GetOk("working_environment_id"); ok {
+		cifs.WorkingEnvironmentID = v.(string)
+	} else {
+		workingEnvDetail, err := client.getWorkingEnvironmentDetail(d, clientID)
+		if err != nil {
+			return err
+		}
+		cifs.WorkingEnvironmentID = workingEnvDetail.PublicID
 	}
 
-	workingEnvDetail, err := client.getWorkingEnvironmentDetail(d, clientID)
-	if err != nil {
-		return nil
+	if v, ok := d.GetOk("svm_name"); ok {
+		cifs.SvmName = v.(string)
 	}
-	cifs.WorkingEnvironmentID = workingEnvDetail.PublicID
-	cifs.WorkingEnvironmentType = workingEnvDetail.WorkingEnvironmentType
-	cifs.SvmName = workingEnvDetail.SvmName
-
 	res, err := client.getCIFS(cifs, clientID)
 	if err != nil {
 		log.Print("Error reading cifs")
 		return err
 	}
 	for _, cifsConfig := range res {
-		if _, ok := d.GetOk("domain"); ok {
-			d.Set("domain", cifsConfig.Domain)
-		}
-		if _, ok := d.GetOk("dns_domain"); ok {
-			d.Set("dns_domain", cifsConfig.DNSDomain)
-		}
-		if _, ok := d.GetOk("ip_addresses"); ok {
+		log.Printf("cifs config: %v", cifsConfig)
+		if cifsConfig.Domain == d.Get("domain").(string) && cifsConfig.DNSDomain == d.Get("dns_domain").(string) &&
+			cifsConfig.NetBIOS == d.Get("netbios").(string) && cifsConfig.OrganizationalUnit == d.Get("organizational_unit").(string) {
 			d.Set("ip_addresses", cifsConfig.IPAddresses)
+			return nil
 		}
-		if _, ok := d.GetOk("netbios"); ok {
-			d.Set("netbios", cifsConfig.NetBIOS)
-		}
-		if _, ok := d.GetOk("organizational_unit"); ok {
-			d.Set("organizational_unit", cifsConfig.OrganizationalUnit)
-		}
-		return nil
-
 	}
-	return fmt.Errorf("Error reading cifs: cifs doesn't exist")
+	return fmt.Errorf("error reading cifs: cifs doesn't exist")
 }
 
 func resourceCVOCIFSDelete(d *schema.ResourceData, meta interface{}) error {
 	log.Printf("Deleting cifs: %#v", d)
 	client := meta.(*Client)
 	clientID := d.Get("client_id").(string)
-	cifs := cifsRequest{}
+	cifs := cifsDeleteRequest{}
 
-	workingEnvDetail, err := client.getWorkingEnvironmentDetail(d, clientID)
-	if err != nil {
-		return nil
+	var workingEnvironmentID string
+	if v, ok := d.GetOk("working_environment_id"); ok {
+		workingEnvironmentID = v.(string)
+	} else {
+		workingEnvDetail, err := client.getWorkingEnvironmentDetail(d, clientID)
+		if err != nil {
+			return err
+		}
+		workingEnvironmentID = workingEnvDetail.PublicID
 	}
-	cifs.WorkingEnvironmentID = workingEnvDetail.PublicID
-	cifs.WorkingEnvironmentType = workingEnvDetail.WorkingEnvironmentType
-	cifs.SvmName = workingEnvDetail.SvmName
-
-	err = client.deleteCIFS(cifs, clientID)
+	cifs.Username = d.Get("username").(string)
+	cifs.Password = d.Get("password").(string)
+	if v, ok := d.GetOk("svm_name"); ok {
+		cifs.SvmName = v.(string)
+	}
+	err := client.deleteCIFS(cifs, workingEnvironmentID, clientID)
 	if err != nil {
 		log.Print("Error deleting cifs")
 		return err
@@ -235,9 +236,9 @@ func resourceCVOCIFSExists(d *schema.ResourceData, meta interface{}) (bool, erro
 		return false, nil
 	}
 	cifs.WorkingEnvironmentID = workingEnvDetail.PublicID
-	cifs.WorkingEnvironmentType = workingEnvDetail.WorkingEnvironmentType
-	cifs.SvmName = workingEnvDetail.SvmName
-
+	if v, ok := d.GetOk("svm_name"); ok {
+		cifs.SvmName = v.(string)
+	}
 	res, err := client.getCIFS(cifs, clientID)
 	if err != nil {
 		log.Print("Error reading cifs")